Multiplex added(executable version).

+++ DSystem.cpp
@@ -1,6 +1,6 @@
 #include <string>
-#include "DSystem.h"
 #include <stdlib.h>
+#include "DSystem.h"
 #include "Multiplex.h"
 
 namespace DSystem {
@@ -19,8 +19,7 @@ namespace DSystem {
 			add_elem_toTail(NULL);
 	}
 
-	phase_space::phase_space(sUInt argNDim): space((int)argNDim, machine_infinity) {
-		
+	phase_space::phase_space(sUInt argNDim): space(argNDim,machine_infinity) {
 	}
 
 	main_ds::main_ds(std::string name_arg = "Default", sUInt ph_DimSpase_arg = 2, sUInt meb_num_of_elems = 0) : main_ps(ph_DimSpase_arg),MEB(meb_num_of_elems) {
